Adjusted returned data from database

+++ RelativityInspector/RelativityInspector.Web/Models/AuditItem.cs
@@ -7,7 +7,7 @@ namespace RelativityInspector.Web.Models
 {
     public class AuditItem
     {
-        public int AuditID { get; internal set; }
+        public long AuditID { get; internal set; }
         public DateTime LastExecutionDate { get; internal set; }
         public string Name { get; internal set; }
         public string Status { get; internal set; }
